+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Font-lock traced vars.
 * New defcustom, `cider-pprint-fn`, allows you to set the function to use when pretty-printing evaluation results.
+* [#1463](https://github.com/clojure-emacs/cider/issues/1463): Assume that `cider-connect` is invoked from within a project,
+and try to associate the created connection with this project automatically.
 
 ### Changes